Refactored hit points to float

+++ Assets/Scripts/Characters/Enemies/EnemyBird.cs
@@ -346,19 +346,19 @@ public class EnemyBird : Bird, IHitPoints
         }
     }
 
-    public IHitPoints.DamageReturn Damage(int damage, IHitPoints.HitType hitType)
+    public IHitPoints.DamageReturn Damage(float damage, IHitPoints.HitType hitType)
     {
         return DoDamage(damage, hitType);
     }
 
-    protected virtual IHitPoints.DamageReturn DoDamage(int damage, IHitPoints.HitType hitType)
+    protected virtual IHitPoints.DamageReturn DoDamage(float damage, IHitPoints.HitType hitType)
     {
         if (m_invTimer > 0.0f)
             return IHitPoints.DamageReturn.NO_DAMAGE;
-        if (m_hitPoints > 0)
+        if (m_hitPoints > 0.0f)
         {
             m_hitPoints -= damage;
-            if (m_hitPoints <= 0)
+            if (m_hitPoints <= 0.0f)
             {
                 Explode();
                 return IHitPoints.DamageReturn.KILLED;    // I've been killed
